refactor(folio): make the layout engine import-graph canvas-free (P2b)
Move the canvas MeasureProvider implementation out of the engine's import path:
measureProvider.ts now owns the interface, registry, and delegators with a
throwing default; the canvas impl exports installCanvasMeasureProvider(), called
at the React editor composition root and a bun test preload. The layout engine
imports measurement only through the provider seam and pure helpers, so it no
longer transitively pulls canvas/DOM — the @stll/folio/server entry and any
future headless/Rust provider path are canvas-free by construction. Completes
Seam 1 (measurement). No behavior change; full suite green.
